# my-page-api/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: JPro My Page API
  description: API for JPro My Page
  version: 1.0.0
paths:
...
  /booking/{bookingId}:
    delete:
      tags:
        - Booking
      operationId: deleteBooking
      summary: Delete booking
      description: Delete a booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Booking deleted
    patch:
      tags:
        - Booking
      operationId: updateBooking
      summary: Update booking
      description: Update a booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Booking updated
...
  /booking/admin/{bookingId}:
    delete:
      tags:
        - Booking
      operationId: adminDeleteBooking
      summary: Admin delete booking
      description: Admin deletes a booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Booking deleted
        '404':
          description: Booking not found
    patch:
      tags:
        - Booking
      operationId: adminUpdateBooking
      summary: Admin update booking
      description: Admin updates a booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Booking updated
        '400':
          description: Bad request
        '404':
          description: Booking not found
...
  /pendingBooking/pendingBookingWin:
    post:
      tags:
        - PendingBooking
      operationId: pickWinnerPendingBooking
      summary: Pending booking becomes a booking
      description: Converts pending bookings to actual bookings (lottery winner selection)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PendingBookingDTO'
      responses:
        '201':
          description: New booking created
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request
  /pendingBooking/{pendingBookingId}:
    delete:
      tags:
        - PendingBooking
      operationId: deleteMyPendingBooking
      summary: Delete a pending booking
      description: Deletes a pending booking if the user has permission
      parameters:
        - name: pendingBookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pending booking deleted successfully
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Forbidden - User does not have permission to delete this pending booking
        '404':
          description: Pending booking not found
  /pendingBooking/admin/{pendingBookingId}:
    delete:
      tags:
        - PendingBooking
      operationId: adminDeletePendingBooking
      summary: Delete a pending booking (admin only)
      description: Deletes any pending booking (requires admin privileges)
      parameters:
        - name: pendingBookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pending booking deleted successfully
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Pending booking not found
    patch:
      tags:
        - PendingBooking
      operationId: updatePendingBooking
      summary: Update a pending booking
      description: Updates a pending booking by ID
      parameters:
        - name: pendingBookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Pending booking updated
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Pending booking not found
  /pendingBooking/pendingPost:
    post:
      tags:
        - PendingBooking
      operationId: createPendingBooking
      summary: Create a new pending booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePendingBooking'
      responses:
        '201':
          description: Pending booking created successfully
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Forbidden
  /pendingBooking/pendingPostForUser:
    post:
      tags:
        - PendingBooking
      operationId: createPendingBookingForUser
      summary: Admin creates a pending booking for a user
      description: Admin creates a new pending booking on behalf of a specific user
      parameters:
        - name: bookingOwnerName
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePendingBooking'
      responses:
        '201':
          description: Pending booking created successfully
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: User not found
...
  /settings/{settingId}:
    patch:
      tags:
        - Settings
      operationId: updateSetting
      summary: Update a setting (admin only)
      parameters:
        - name: settingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Setting'
      responses:
        '200':
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
        '404':
          description: Setting not found
...
components:
  schemas:
...
    Apartment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cabin_name:
          type: string
        sort_order:
          type: integer
      required:
        - id
        - cabin_name
    BookingUpdate:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        apartmentId:
          type: integer
          format: int64
...
    PendingBookingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        apartment:
          $ref: '#/components/schemas/Apartment'
        employeeName:
          type: string
        createdDate:
          type: string
          format: date
      required:
        - startDate
        - endDate
        - createdDate
...
    Setting:
      type: object
      properties:
        settingId:
          type: string
        settingValue:
          type: string
        settingDescription:
          type: string
      required:
        - settingId
        - settingValue
...
    CreatePendingBooking:
      type: object
      properties:
        apartmentID:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
        - apartmentID
        - startDate
        - endDate
...